Reduces admin api e2e code duplication
